+++ dart/attributes/lib/src/entity/entity.dart
@@ -35,6 +35,21 @@ abstract interface class Entity {
   factory Entity.view({Object? id, required Map<String, Object?> properties}) =
       EntityBase.view;
 
+  /// A new entity from JSON Object containg an optional identifier in "id" and
+  /// the required properties (JSON Object) in "properties".
+  ///
+  /// This is a factory creating an instance of [EntityBase].
+  factory Entity.fromJson(Map<String, Object?> json) = EntityBase.fromJson;
+
+  /// A new entity from JSON Object containg an optional identifier in "id" and
+  /// the required properties (JSON Object) in "properties".
+  ///
+  /// The underlying map is an object tree as parsed by the standard
+  /// `json.decode()` of the `dart:convert` package.
+  ///
+  /// This is a factory creating an instance of [EntityBase].
+  factory Entity.decodeJson(String source) = EntityBase.decodeJson;
+
   /// An empty entity with empty properties and without id.
   ///
   /// This is a factory creating an instance of [EntityBase].
@@ -45,4 +60,19 @@ abstract interface class Entity {
 
   /// The required [properties] for this entity, allowed to be empty.
   DataObject get properties;
+
+  /// Returns data as an encodable object compatible with `json.encode()`.
+  /// 
+  /// The returned JSON Object contains an optional identifier in "id" and
+  /// the required properties (JSON Object) in "properties".
+  Map<String, Object?> toJson();
+
+  /// Encodes this entity into a JSON string containing a JSON Object.
+  /// 
+  /// The returned JSON Object contains an optional identifier in "id" and
+  /// the required properties (JSON Object) in "properties".
+  ///
+  /// Any `DateTime` objects contained are encoded using [encodeTime] if
+  /// provided, or otherwise using `DateTime.toIso8601String()` method.
+  String encodeJson({Object Function(DateTime time)? encodeTime});
 }